+++ packages/build-info/CHANGELOG.md
@@ -64,6 +64,20 @@
   * dependencies
     * @netlify/framework-info bumped from ^9.8.8 to ^9.8.9
 
+## [7.3.0](https://github.com/netlify/build/compare/build-info-v7.2.0...build-info-v7.3.0) (2023-06-30)
+
+
+### Features
+
+* detect lang runtime ([#5105](https://github.com/netlify/build/issues/5105)) ([7242ddd](https://github.com/netlify/build/commit/7242dddef6307f1c673ba94f7026a85eb5a0fb59))
+
+## [7.2.0](https://github.com/netlify/build/compare/build-info-v7.1.1...build-info-v7.2.0) (2023-06-29)
+
+
+### Features
+
+* make events async ([#5106](https://github.com/netlify/build/issues/5106)) ([6fa5220](https://github.com/netlify/build/commit/6fa5220bc6d9f9e390564cd2f1a3a4c5bc9cd206))
+
 ## [7.1.1](https://github.com/netlify/build/compare/build-info-v7.1.0...build-info-v7.1.1) (2023-06-29)